requirements.txt
langchain==0.1.19
python-binance==1.0.19
supervisor==4.2.5
pytest==8.4.2
anyio==4.11.0
mcp==1.0.0
requests==2.31.0

# Additional dependencies for enhanced features
numpy==1.26.4
pandas==2.2.2
ta==0.11.0
tabulate==0.9.0
pytest-asyncio==0.24.0
toml==0.10.2

# Web UI dependencies
streamlit==1.35.0
plotly==5.20.0

# Optional: For enhanced monitoring (uncomment if needed)
# prometheus-client==0.20.0
